requirements.txt
gradio>=4.44,<5
requests>=2.31.0
python-dotenv>=1.0.1
